requirements.txt
click==8.1.7
rich==13.7.1
cryptography==43.0.1
pycryptodomex==3.20.0
watchdog==4.0.1
pandas==2.2.2
numpy==2.1.1
scikit-learn==1.5.2
jinja2==3.1.4
reportlab==4.2.2
pyyaml==6.0.2
python-dateutil==2.9.0.post0
psutil==6.0.0
flask==3.0.0
scapy==2.5.0
geoip2==4.7.0
argon2-cffi==23.1.0
reportlab==4.2.2
waitress==3.0.0